refactor(A1): split DatabaseManager God Class into repository subpackage
将 1297 行的 DatabaseManager（41 个方法、7 项职责）拆分为 app/core/database/
子包，采用 Repository 模式 + Facade 保持完全向后兼容：

- connection.py: DatabaseConnection（连接管理、线程锁、schema 迁移）
- sync_records.py: SyncRecordsRepository（6 个方法）
- trakt.py: TraktRepository（8 个方法）
- feiniu.py: FeiniuRepository（8 个方法）
- inbox.py: InboxRepository（11 个方法，含 backfill_historical_error_notifications）
- __init__.py: DatabaseManager facade + database_manager 单例

跨仓库依赖通过构造器注入：inbox 依赖 feiniu（backfill 读 feiniu_meta），
sync_records 依赖 inbox（update_sync_record_status 标记通知已读）。
__init__.py 重新导出 sqlite3 / logger / 两个元数据常量，保证测试 patch
（如 app.core.database.sqlite3.connect）继续生效。

1894 个测试全部通过，ruff check/format 全部通过。
